// Client/Redux/product/productType.ts
export enum ProductType {
  SET_PRODUCTS = "SET_PRODUCTS",
  ADD_TO_CART = "ADD_TO_CART",
  ADD_TOTALS = "ADD_TOTALS",
  SET_ITEM_CART = "SET_ITEM_CART",
  GET_ITEM_CART = "GET_ITEM_CART",

  SET_SINGLE_PRODUCT = "SET_SINGLE_PRODUCT",
  GET_SINGLE_PRODUCT = "GET_SINGLE_PRODUCT",

  INCREMENT = "INCREMENT",
  DECREMENT = "DECREMENT",
  REMOVE = "REMOVE",
  CLEAR = "CLEAR",
}